+++ src/entity/burger.rs
@@ -17,10 +17,10 @@ impl Burger
         }
     }
 
-    pub fn get_id(&self) -> u8 {
+    pub fn get_id(&self) -> u32 {
         self.id
     }
-    pub fn set_id(&mut self, id: u8) {
+    pub fn set_id(&mut self, id: u32) {
         self.id = id;
     }
