+++ src/entity/ingredient.rs
@@ -15,10 +15,10 @@ impl Ingredient
         }
     }
 
-    pub fn get_id(&self) -> u8 {
+    pub fn get_id(&self) -> u32 {
         self.id
     }
-    pub fn set_id(&mut self, id: u8) {
+    pub fn set_id(&mut self, id: u32) {
         self.id = id;
     }
